+++ src/main/java/za/ac/cput/service/IAdminService.java
@@ -3,8 +3,9 @@ package za.ac.cput.service;
 import za.ac.cput.domain.Admin;
 import java.util.List;
 
-public interface IAdminService extends IService<Admin, Long> {
+public interface IAdminService extends IService<Admin, Integer> {
     List<Admin> getAll();
 
-    void delete(Long adminID);
+
+    void delete(Integer adminID);
 }
